backend/user-service/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.2
python-dotenv==1.0.1
Werkzeug==3.1.3
pillow==11.0.0
PyJWT==2.10.1
requests==2.32.3
ldap3==2.9.1
flask-jwt-extended=4.7.1